Move TargetID into separate include/, src/ files
